feat: split analyzer passes
